+++ app/src/test/java/woowacourse/shopping/coupon/CouponViewModelTest.kt
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.CoroutinesTestExtension
 import woowacourse.shopping.InstantTaskExecutorExtension
-import woowacourse.shopping.cartItems
+import woowacourse.shopping.cartItemsBySize
 import woowacourse.shopping.domain.model.CartItem
 import woowacourse.shopping.domain.model.Coupon
 import woowacourse.shopping.domain.repository.CartRepository
@@ -41,7 +41,7 @@ class CouponViewModelTest {
     fun `적용할 수 있는 쿠폰을 로드한다`(): Unit =
         runBlocking {
             // given
-            val cartItems = cartItems(5)
+            val cartItems = cartItemsBySize(5)
             val selectedCartItemIds = cartItems.map { it.id }
             val coupons = List(5) { fixedCoupon(id = it, discount = 1_000) }
             setUpViewModel(selectedCartItemIds, cartItems, coupons)
@@ -58,7 +58,7 @@ class CouponViewModelTest {
     fun `적용할 수 있는 쿠폰이 없는 경우 쿠폰 목록이 비어있다`(): Unit =
         runBlocking {
             // given
-            val cartItems = cartItems(5)
+            val cartItems = cartItemsBySize(5)
             val selectedCartItemIds = cartItems.map { it.id }
             setUpViewModel(selectedCartItemIds, cartItems, listOf())
 
@@ -74,7 +74,7 @@ class CouponViewModelTest {
     fun `장바구니 상품을 결제한다`(): Unit =
         runBlocking {
             // given
-            val cartItems = cartItems(5)
+            val cartItems = cartItemsBySize(5)
             val selectedCartItemIds = cartItems.map { it.id }
             setUpViewModel(selectedCartItemIds, cartItems, listOf())
